create BaseAnalysis class
